+++ backend/src/main/java/com/meong9/backend/domain/member/service/KakaoService.java
@@ -10,7 +10,7 @@ import com.meong9.backend.domain.member.entity.Member;
 import com.meong9.backend.domain.member.repository.MemberRepository;
 import com.meong9.backend.global.auth.entity.MemberDetails;
 import com.meong9.backend.global.auth.refreshtoken.RefreshTokenService;
-import com.meong9.backend.global.auth.utils.JwtProvider;
+import com.meong9.backend.global.auth.jwt.JwtProvider;
 import com.meong9.backend.global.mediafile.entity.FileType;
 import com.meong9.backend.global.mediafile.dto.ImageMetadataDto;
 import com.meong9.backend.global.mediafile.dto.S3UploadResultDto;
@@ -19,7 +19,6 @@ import com.meong9.backend.global.exception.AuthenticationException;
 import com.meong9.backend.global.exception.ConflictException;
 import com.meong9.backend.global.mediafile.repository.MediaFileRepository;
 import com.meong9.backend.global.mediafile.service.MediaFileService;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
